Add files via upload
testblackjack_stim_draw.py : This contains test script to run with python inbuilt unit test module for unit-testing. Expected return from the function is 0 for the draw condition between the player and the dealer. 

testblackjack_stim_loose.py : This contains test script to run with python inbuilt unit test module for unit-testing. Expected return from the function is -1 for the loose condition of the player.

testblackjack_stim_win.py : This contains test script to run with python inbuilt unit test module for unit-testing. Expected return from the function is 1 for the win situation of the player.